% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_color_vector}
\alias{create_color_vector}
\title{Wrap color vector in function.}
\usage{
create_color_vector(colors = statworx_highlights, ...)
}
\arguments{
\item{colors}{List of colors}

\item{...}{Name of color from "statworx_highlights"}
}
\value{
Hex code of the color
}
\description{
Wrap color vector in function.
}
