% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{statworx_flip}
\alias{statworx_flip}
\title{statworx Theme based on theme_classic}
\usage{
statworx_flip(font = "Arial", dark = FALSE)
}
\arguments{
\item{font}{set default font}

\item{dark}{choose dark mode}
}
\description{
Sets a pre-defined theme as the standard ggplot theme via
\code{\link[ggplot2:ggtheme]{theme_classic}}
}
\examples{
\dontrun{
library(statworxTheme)
statworx_flip()
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point()
}
}
\keyword{theme}
