% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{statworx_highlights}
\alias{statworx_highlights}
\title{statworx palettes}
\format{
An object of class \code{character} of length 10.
}
\usage{
statworx_highlights
}
\description{
statworx palettes
}
\keyword{datasets}
