% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{statworx_palettes}
\alias{statworx_palettes}
\title{statworx palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
statworx_palettes
}
\description{
Vector of color codes.
}
\keyword{datasets}
